+++ tests/test_classes.py
@@ -5,9 +5,9 @@ import unittest
 from itertools import count
 from typing import Optional, List, Union
 
-from rdflib.namespace import XSD
 import pydantic
 import rdflib
+import yaml
 from pydantic import EmailStr, model_validator
 from pydantic import ValidationError
 from pydantic import field_validator, Field
@@ -15,6 +15,7 @@ from rdflib.plugins.shared.jsonld.context import Context
 from typing_extensions import Annotated
 
 import ontolutils
+from ontolutils import SCHEMA
 from ontolutils import Thing, urirefs, namespaces, build, Property
 from ontolutils import as_id
 from ontolutils import get_urirefs, get_namespaces, set_config
@@ -228,21 +229,60 @@ class TestNamespaces(unittest.TestCase):
         thing2 = Thing(label='Thing 2', id='https://example.com/thing2')
         self.assertTrue(thing1 < thing2)
 
-    def test_language(self):
-
-        # with self.assertRaises(ValueError):
-        #     LangString(value='invalid', lang='en', datatype=XSD.date)
-
-        thing_en = Thing(label=LangString(value='a thing', lang='en'))
+    def test_language_string0(self):
+        thing_en = Thing(label=rdflib.Literal(lexical_or_value='a thing.', lang='en'))
         ttl = thing_en.model_dump_ttl()
         self.assertEqual(ttl, """@prefix owl: <http://www.w3.org/2002/07/owl#> .
 @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
 
 [] a owl:Thing ;
-    rdfs:label "a thing"@en .
+    rdfs:label "a thing."@en .
 
 """)
 
+    def test_language_string1(self):
+        self.assertNotEqual("x", LangString(value='upward', lang=None))
+        self.assertEqual("a thing", LangString(value="a thing", lang="en").value)
+        self.assertEqual("a thing", LangString(value="a thing").value)
+        self.assertEqual("a thing", LangString(value="a thing"))
+        self.assertEqual("a thing", LangString(value="a thing", lang="en"))
+        self.assertNotEqual("a thing", LangString(value="another thing").value)
+        self.assertNotEqual("a thing", LangString(value="another thing", lang="en").value)
+
+        self.assertNotEqual(LangString(value="Hello", lang="en"), "Hello@fr")
+        thing_en = Thing(label=LangString(value='a thing.', lang='en'))
+        with set_config(show_lang_in_str=True):
+            ttl = thing_en.model_dump_ttl()
+            self.assertEqual(ttl, """@prefix owl: <http://www.w3.org/2002/07/owl#> .
+@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
+
+[] a owl:Thing ;
+    rdfs:label "a thing."@en .
+
+""")
+        with set_config(show_lang_in_str=False):
+            self.assertEqual(str(LangString(value="a thing", lang="en")), "a thing")
+        with set_config(show_lang_in_str=True):
+            self.assertEqual(str(LangString(value="a thing", lang="en")), "a thing@en")
+        with set_config(show_lang_in_str=True):
+            self.assertEqual(str(LangString(value="a thing")), "a thing")
+
+    def test_language_string2(self):
+        thing_en = Thing(
+            label=[LangString(value='a thing.', lang='en'),
+                   LangString(value='ein Ding.', lang='de'), ]
+        )
+        ttl = thing_en.model_dump_ttl()
+        self.assertEqual(ttl, """@prefix owl: <http://www.w3.org/2002/07/owl#> .
+@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
+
+[] a owl:Thing ;
+    rdfs:label "ein Ding."@de,
+        "a thing."@en .
+
+""")
+
+    def test_language_string3(self):
         thing_en = Thing(label="deutsch@de")
         ttl = thing_en.model_dump_ttl()
         self.assertEqual(ttl, """@prefix owl: <http://www.w3.org/2002/07/owl#> .
@@ -253,30 +293,69 @@ class TestNamespaces(unittest.TestCase):
 
 """)
 
-#         thing_en = Thing(label=LangString(value='2025-01-01', datatype=XSD.date))
-#         ttl = thing_en.model_dump_ttl()
-#         self.assertEqual(ttl, """@prefix owl: <http://www.w3.org/2002/07/owl#> .
-# @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
-# @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-#
-# [] a owl:Thing ;
-#     rdfs:label "2025-01-01"^^xsd:date .
-#
-# """)
+        #         thing_en = Thing(label=LangString(value='2025-01-01', datatype=XSD.date))
+        #         ttl = thing_en.model_dump_ttl()
+        #         self.assertEqual(ttl, """@prefix owl: <http://www.w3.org/2002/07/owl#> .
+        # @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
+        # @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+        #
+        # [] a owl:Thing ;
+        #     rdfs:label "2025-01-01"^^xsd:date .
+        #
+        # """)
 
-        thing_en = Thing(label='2025-01-01')
+    def test_language_string4(self):
+        thing_en = Thing(label=[
+            "a thing.@en",
+            "ein Ding@de"
+        ])
         ttl = thing_en.model_dump_ttl()
         self.assertEqual(ttl, """@prefix owl: <http://www.w3.org/2002/07/owl#> .
 @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
 
 [] a owl:Thing ;
-    rdfs:label "2025-01-01" .
+    rdfs:label "ein Ding"@de,
+        "a thing."@en .
 
 """)
+
+    def test_lang_string_yaml_support(self):
+        ls = LangString(value="hallo", lang="de")
+        with open("output.yaml", "w", encoding="utf-8") as f:
+            yaml.dump(ls, f, allow_unicode=True)
+        with open("output.yaml", "r", encoding="utf-8") as f:
+            ls2 = yaml.load(f, Loader=yaml.Loader)
+        self.assertEqual(ls, ls2)
 
     def test__repr_html_(self):
         thing = Thing(label='Thing 1')
         self.assertEqual(thing._repr_html_(), f'Thing(id={thing.id}, label=Thing 1)')
+
+    def test_serialize_date(self):
+        @namespaces(dcterms="http://purl.org/dc/terms/")
+        @urirefs(MyThing='foaf:MyThing',
+                 created='dcterms:created')
+        class MyThing(Thing):
+            created: datetime.datetime = None
+
+        mything = MyThing(created=datetime.datetime(year=2025, month=10, day=1, hour=12, minute=30))
+        ttl = mything.serialize("ttl")
+        self.assertEqual(ttl, """@prefix dcterms: <http://purl.org/dc/terms/> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+
+[] a <foaf:MyThing> ;
+    dcterms:created "2025-10-01T12:30:00"^^xsd:dateTime .
+
+""")
+        mything = MyThing(created=datetime.datetime(year=2025, month=10, day=1))
+        ttl = mything.serialize("ttl")
+        self.assertEqual(ttl, """@prefix dcterms: <http://purl.org/dc/terms/> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+
+[] a <foaf:MyThing> ;
+    dcterms:created "2025-10-01"^^xsd:date .
+
+""")
 
     def test_thing_get_jsonld_dict(self):
         with self.assertRaises(TypeError):
@@ -300,6 +379,7 @@ class TestNamespaces(unittest.TestCase):
             thing_dict['@context'],
             {'owl': 'http://www.w3.org/2002/07/owl#',
              'rdfs': 'http://www.w3.org/2000/01/rdf-schema#',
+             'schema': 'https://schema.org/',
              'skos': 'http://www.w3.org/2004/02/skos/core#',
              'dcterms': 'http://purl.org/dc/terms/'
              }
@@ -532,7 +612,7 @@ class TestNamespaces(unittest.TestCase):
                  affiliation='schema:affiliation')
         class Person(Agent):
             firstName: str = None
-            affiliation: Organization = None
+            affiliation: Union[str, Organization] = None
 
         person = Person(
             label='Person 1',
@@ -547,6 +627,13 @@ class TestNamespaces(unittest.TestCase):
         self.assertEqual(jsonld_dict['schema:affiliation']['rdfs:label'], 'Organization 1')
         self.assertEqual(jsonld_dict['rdfs:label'], 'Person 1')
         self.assertEqual(jsonld_dict['@type'], 'foaf:Person')
+
+        person = Person(
+            label='Person 1',
+            affiliation='schema:Organization/123',
+        )
+
+        print(person.serialize("ttl"))
 
     def test_prov(self):
         @namespaces(prov="https://www.w3.org/ns/prov#",
@@ -605,6 +692,7 @@ class TestNamespaces(unittest.TestCase):
                 "prov": "https://www.w3.org/ns/prov#",
                 "rdfs": "http://www.w3.org/2000/01/rdf-schema#",
                 "dcterms": "http://purl.org/dc/terms/",
+                'schema': 'https://schema.org/',
                 "skos": "http://www.w3.org/2004/02/skos/core#",
                 "foaf": "http://xmlns.com/foaf/0.1/",
             },
@@ -637,6 +725,7 @@ class TestNamespaces(unittest.TestCase):
                 "owl": "http://www.w3.org/2002/07/owl#",
                 "prov": "https://www.w3.org/ns/prov#",
                 "rdfs": "http://www.w3.org/2000/01/rdf-schema#",
+                'schema': 'https://schema.org/',
                 "dcterms": "http://purl.org/dc/terms/",
                 "skos": "http://www.w3.org/2004/02/skos/core#",
                 "foaf": "http://xmlns.com/foaf/0.1/",
@@ -764,10 +853,11 @@ class TestNamespaces(unittest.TestCase):
         self.assertDictEqual(mt.urirefs, get_urirefs(Thing))
         self.assertDictEqual(mt.urirefs,
                              {'Thing': 'owl:Thing', 'closeMatch': 'skos:closeMatch', 'exactMatch': 'skos:exactMatch',
-                              'label': 'rdfs:label', 'relation': 'dcterms:relation'})
+                              'label': 'rdfs:label', 'about': 'schema:about', 'relation': 'dcterms:relation'})
         self.assertDictEqual(mt.namespaces, get_namespaces(Thing))
         self.assertDictEqual(mt.namespaces, {'owl': 'http://www.w3.org/2002/07/owl#',
                                              'rdfs': 'http://www.w3.org/2000/01/rdf-schema#',
+                                             'schema': 'https://schema.org/',
                                              'skos': 'http://www.w3.org/2004/02/skos/core#',
                                              'dcterms': 'http://purl.org/dc/terms/'})
 
@@ -785,6 +875,7 @@ class TestNamespaces(unittest.TestCase):
             "@context": {
                 "owl": "http://www.w3.org/2002/07/owl#",
                 "rdfs": "http://www.w3.org/2000/01/rdf-schema#",
+                'schema': 'https://schema.org/',
                 "dcterms": "http://purl.org/dc/terms/",
                 "skos": "http://www.w3.org/2004/02/skos/core#",
                 "foaf": "http://xmlns.com/foaf/0.1/"
@@ -1102,3 +1193,61 @@ agents:123 a foaf:Agent ;
         org1 = Organization(members=p1)
         org2 = Organization(members=p2.map(Person1))
         # org3 = Organization(members=[p1, p2])
+
+    def test_about(self):
+        @namespaces(foaf="http://xmlns.com/foaf/0.1/")
+        @urirefs(Agent='foaf:Agent',
+                 name='foaf:lastName')
+        class Agent(Thing):
+            """Pydantic Model for http://xmlns.com/foaf/0.1/Agent
+            Parameters
+            ----------
+            mbox: EmailStr = None
+                Email address (foaf:mbox)
+            """
+            name: str = Field(default=None, alias="lastName")  # name is synonymous to lastName
+
+        #         with self.assertRaises(ValidationError):
+        #             Agent(
+        #                 name="John Doe",
+        #                 about="A person"
+        #             )
+        #         p = Agent(
+        #             name="John Doe",
+        #             about="http://example.org/123"
+        #         )
+        #         p = Agent(
+        #             name="John Doe",
+        #             about=["http://example.org/123", "http://example.org/456"]
+        #         )
+        #         p = Agent(
+        #             name="John Doe",
+        #             about=["http://example.org/123", Thing(id="http://example.org/456")]
+        #         )
+        #         ttl = p.model_dump_ttl()
+        #         self.assertEqual(ttl, """@prefix foaf: <http://xmlns.com/foaf/0.1/> .
+        # @prefix owl: <http://www.w3.org/2002/07/owl#> .
+        # @prefix schema: <https://schema.org/> .
+        #
+        # <http://example.org/456> a owl:Thing .
+        #
+        # [] a foaf:Agent ;
+        #     foaf:lastName "John Doe" ;
+        #     schema:about <http://example.org/456>,
+        #         "http://example.org/123" .
+        #
+        # """)
+        p = Agent(
+            name="John Doe",
+            about=SCHEMA.about
+        )
+        ttl = p.model_dump_ttl()
+        print(ttl)
+        self.assertEqual(ttl, """@prefix foaf: <http://xmlns.com/foaf/0.1/> .
+@prefix schema: <https://schema.org/> .
+
+[] a foaf:Agent ;
+    foaf:lastName "John Doe" ;
+    schema:about schema:about .
+
+""")